requirements.txt
aiogram==3.18.0
openai==1.57.0
python-dotenv==1.0.0